+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 0.4.0
-- Implemented required() in include syntax
+- Implemented required() function in include syntax
+- Implemented file() function in include syntax
 
 ## 0.3.2
 - Fixed UnresolvedSubstitution hashing